refactor amp/bento-date-countdown
